requirements.txt
pandas==2.0.3
scikit-learn==1.2.2
PyYAML==6.0.2
numpy==1.24.0
